refactor: move struct operations to a separate module
